OSIS-3851 moved EducationGroupYearHyperlinkedIdentityField from utils to education_group
